+++ lib/src/models/list_place/list_item_place_model.dart
