+++ lib/src/resources/list_place_provider.dart
@@ -1,7 +1,7 @@
 import 'package:http/http.dart' show Client;
 import 'dart:convert';
 import '../data/constants.dart';
-import '../models/list_item_place_model.dart';
+import '../models/list_place/list_item_place_model.dart';
 
 class ListPlaceProvider {
