+++ lib/src/screens/list_place.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import '../blocs/list_place_provider.dart';
-import '../models/list_item_place_model.dart';
-import '../models/item_place_model.dart';
+import '../models/list_place/list_item_place_model.dart';
+import '../models/list_place/item_place_model.dart';
 import '../widgets/item_place.dart';
 
 class ListPlace extends StatelessWidget {
