+++ lib/src/screens/list_place_detail.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import '../models/item_place_model.dart';
+import '../models/list_place/item_place_model.dart';
 import '../data/dimens.dart';
 import '../data/constants.dart';
